fix ppc/powerpc arch configuration ambiguity
The configure script does not properly determine the real arch of the
kernel used for PowerPC, either "ppc" or "powerpc". This patch updates
configure to look for the string "ppc" in the symbolic link
"include/asm" of the kernel. If it does find it, the arch "ppc" is
selected, "powerpc" otherwise.

Signed-off-by: Wolfgang Grandegger <wg@denx.de>
Tested-by: Giammarco Zacheo <g.zacheo@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@web.de>