rtcfg: Prevent usage of freed rtskb data
Prevent usage of freed rtskb 'stage_1_cfg' by using 'return' rather than
'break' in the error condition where the stage 1 cfg is invalid.  Since
the call to kfree_rtskb() also cleans up 'rtskb->data' the data pointed
to by 'stage_1_cfg' is no longer allocated.  This is also consistent
with the other error conditions in this part of the code.

Signed-off-by: James Kilts <jameskilts@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@web.de>