RT-TCP: Fix race during retransmit timer requeuing
timerwheel_add_timer called from within the timer handler raced with
timerwheel_pivot on the timer->slot. The latter used to overwrite the
slot valued set while requeuing the timer.

Fix this by checking the timer state atomically after the handler
returned, only updating it if it is still TIMERWHEEL_TIMER_TRIGGERED.
This is safe as the transition to TIMERWHEEL_TIMER_TRIGGERED can only
take place in the timerwheel task itself.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>