RT-TCP: Use refcounter to protect timer during handler call
When restarting a timer from its handler or before the botton-half of
the timerwheel_pivot loop ran, we loose the information that the loop
has still a reference on this object. Thus, timerwheel_remove_timer_sync
may return before the object was actually dropped by the pivot loop and
the latter could later on dereference an invalid pointer.

Fix this by using a reference counter to track the usage between
timerwheel_get_from_current_slot and the return from the handler. We do
not need an atomic counter as only the timerwheel tasks modifies it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>