Fix copying and setting of MAC addresses
Since kernel 2.6.31 dev_addr in struct net_device is a pointer and
not an array itself. Memcpying or memsetting the MAC address with
a sizeof(dev_addr) as length argument does not work properly as
now the length of the pointer is taken and not the size of the
array.

This patch sets the memcpy and memset lengths in three files to
MAX_ADDR_LEN. This define exists in the kernel since the old 2.4
days and didn't change its meaning, so it can be considered as a
safe length to use for these memcpy and memset operations.

Signed-off-by: Sebastian Smolorz <smolorz@rts.uni-hannover.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>