Add support for pre-mapping of rtskbs
When RTnet is executed on a system with IOMMU enabled, dma_map/unmap
services will no longer be lightweight and, instead, call into the IOMMU
core with all sorts of Linux locks.

To avoid having to call those helpers over RT code path, this introduces
a way for NIC driver to pre-map all rtskbs to all devices (that support
pre-mapping). Two device callbacks are introduced. One is invoked when a
new rtskb is added to some pool, the other when it is destroyed.
Moreover, all rtskbs are tracked, and when a device that supports pre-
mapping is registered, already existing rtskbs are presented to its
callback as well.

The device interface probably needs some refactoring to simplify the
common case of PCI devices, but let's give this a start first.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>